refactoring, Deleted Code, Comments
