+++ backend/src/index.js
@@ -1,12 +1,12 @@
 /* eslint-disable no-console */
 import { ApolloServer } from 'apollo-server';
 // import datasource from './main/datasource';
-import InMemoryDataSource from './main/InMemoryDataSource';
+import ResolverHandler from './ResolverHandler';
 import context from './context';
 import Schema from './schema';
 
 // const { users, posts } = datasource;
-const db = new InMemoryDataSource();
+const db = new ResolverHandler();
 const dataSources = () => ({ db });
 // const context = ({ req, res }) => ({ req, res });
 const opts = {
