+++ backend/src/resolver.js
@@ -15,10 +15,9 @@ export default ([{ schema, executor }]) => ({
   },
 
   Mutation: {
-    write:
-    (parent, args, context, info) => context.dataSources.db.createPost(args, context, info, schema),
-    delete: (parent, args, context, info) => context.dataSources.db.deletePost(args, context, info, schema),
-    upvote: (parent, args, context, info) => context.dataSources.db.upvotePost1(args, context, info, schema),
+    write: (parent, args, context, info) => context.dataSources.db.createPost(args, context, info, schema),
+    delete: (parent, args, context, info) => context.dataSources.db.deletePost(args, context, executor),
+    upvote: (parent, args, context, info) => context.dataSources.db.upvotePost(args, context, executor),
     downvote: (parent, args, context) => context.dataSources.db.downvotePost(args, context, executor, schema),
     signup: (parent, args, context) => context.dataSources.db.createUser(args, context, executor),
     login: (parent, args, context) => context.dataSources.db.loginUser(args, context, executor),
